+++ src/core/Controller.h
@@ -1,12 +1,6 @@
 /* XMRig
- * Copyright 2010      Jeff Garzik <jgarzik@pobox.com>
- * Copyright 2012-2014 pooler      <pooler@litecoinpool.org>
- * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
- * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
- * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
- * Copyright 2018-2019 SChernykh   <https://github.com/SChernykh>
- * Copyright 2016-2019 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright 2018-2021 SChernykh   <https://github.com/SChernykh>
+ * Copyright 2016-2021 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -27,12 +21,15 @@
 
 
 #include "base/kernel/Base.h"
-#include "base/tools/Object.h"
+
+
+#include <memory>
 
 
 namespace xmrig {
 
 
+class HwApi;
 class Job;
 class Miner;
 class Network;
@@ -58,8 +55,12 @@ public:
     void execCommand(char command);
 
 private:
-    Miner *m_miner     = nullptr;
-    Network *m_network = nullptr;
+    std::shared_ptr<Miner> m_miner;
+    std::shared_ptr<Network> m_network;
+
+#   ifdef XMRIG_FEATURE_API
+    std::shared_ptr<HwApi> m_hwApi;
+#   endif
 };