+++ src/hw/api/api.cmake
@@ -1,0 +1,11 @@
+if (WITH_HTTP)
+    add_definitions(/DXMRIG_FEATURE_DMI)
+
+    list(APPEND HEADERS
+        src/hw/api/HwApi.h
+        )
+
+    list(APPEND SOURCES
+        src/hw/api/HwApi.cpp
+        )
+endif()