+++ src/hw/msr/MsrItem.h
@@ -1,14 +1,6 @@
 /* XMRig
- * Copyright 2010      Jeff Garzik <jgarzik@pobox.com>
- * Copyright 2012-2014 pooler      <pooler@litecoinpool.org>
- * Copyright 2014      Lucas Jones <https://github.com/lucasjones>
- * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
- * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
- * Copyright 2017-2019 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
- * Copyright 2018      Lee Clagett <https://github.com/vtnerd>
- * Copyright 2018-2019 tevador     <tevador@gmail.com>
- * Copyright 2018-2020 SChernykh   <https://github.com/SChernykh>
- * Copyright 2016-2020 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright (c) 2018-2021 SChernykh   <https://github.com/SChernykh>
+ * Copyright (c) 2016-2021 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -39,9 +31,6 @@ namespace xmrig
 {
 
 
-class RxDataset;
-
-
 class MsrItem
 {
 public:
@@ -56,6 +45,11 @@ public:
     inline uint32_t reg() const     { return m_reg; }
     inline uint64_t value() const   { return m_value; }
     inline uint64_t mask() const    { return m_mask; }
+
+    static inline uint64_t maskedValue(uint64_t old_value, uint64_t new_value, uint64_t mask)
+    {
+        return (new_value & mask) | (old_value & ~mask);
+    }
 
     rapidjson::Value toJSON(rapidjson::Document &doc) const;
     String toString() const;